Crate quilkin_corrosion

Re-exports

pub use corro_api_types as api;

pub use corro_types as types;

Modules

codec

db

metrics

persistent

Implementation for a persistent connection between a client (mutating agent or subscribing proxy) and server (relay).

pubsub

Reimplementation of some pieces of https://github.com/superfly/corrosion/blob/main/crates/corro-agent/src/api/public/pubsub.rs

schema

Structs

Path

A slice of a UTF-8 path (akin to str).

PathBuf

An owned, mutable UTF-8 path (akin to String).

SmallVec

A Vec-like container that can store a small number of elements inline.

Tripwire

A Future that completes once the program is requested to shutdown. This gives a chance for other Futures to do additional cleanup before being dropped.

Functions

ip_to_peer

Type Aliases

Peer