quilkin_corrosion::gossip::changes

Function handle_changes

pub async fn handle_changes(
    ctx: ChangeCtx,
    opts: ChangeOptions,
    changes_rx: Receiver<Change>,
    tripwire: Tripwire,
)

Bundle incoming changes to optimise transaction sizes with SQLite

Performance tradeoff: introduce latency (with a max timeout) to apply changes more efficiently.

This function used by broadcast receivers and sync receivers